requirements.txt
attrdict
PsychoPy>=2023.1.0
gym==0.25.2
gym-retro
pandas>=1.1.1
python-dotenv
tqdm>=4.60.0
textdistance
colorama
psychopy-sounddevice
termcolor
pyaudio
